chore: move all postprocess functions to one file
