--- src/utils/postProcessService.ts
@@ -1,13 +1,0 @@
-import type { Service } from '../types/client';
-import { postProcessServiceImports } from './postProcessServiceImports';
-import { postProcessServiceOperations } from './postProcessServiceOperations';
-
-export const postProcessService = (service: Service): Service => {
-    const clone = { ...service };
-    clone.operations = postProcessServiceOperations(clone);
-    clone.operations.forEach(operation => {
-        clone.imports.push(...operation.imports);
-    });
-    clone.imports = postProcessServiceImports(clone);
-    return clone;
-};
